requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
yfinance==0.2.28
xgboost==1.7.6
mlflow==2.8.1
evidently==0.4.12
pydantic==1.10.13
streamlit==1.28.1
matplotlib==3.7.2
plotly==5.17.0
statsmodels==0.14.0
textblob==0.17.1
nltk==3.8.1
python-dotenv==1.0.0
joblib==1.3.2
requests==2.31.0
pytz==2023.3
loguru==0.7.2
fastapi==0.104.1
uvicorn==0.24.0
